Factors out NN
